Finished writing _eval functions
