+++ sympy/matrices/dense.py
@@ -402,20 +402,8 @@ class DenseMatrix(CommonMatrix, MatrixBase):
         cols = self.cols
         indices = (i*cols + j for i in rowsList for j in colsList)
         return self._new(len(rowsList), len(colsList), list(mat[i] for i in indices), copy=False)
-    #def _eval_trace(self):
-    #    return MatrixBase.trace(self)
-    #def _eval_inverse(self, method=None, **kwargs):
-    #    return MatrixBase.inv(self, method, **kwargs)
-    def _eval_inverse_GE(self, iszerofunc=_iszero):
-        aug = self.row_join(Matrix.eye(self.rows))
-        aug_rref = aug.rref(iszerofunc=_iszero, simplify=True)[0]
-        if any(iszerofunc(aug_rref[i,i]) for i in range(aug_rref.rows)):
-            raise ValueError("Matrix det == 0; not invertible.")
-        return self._new(self.rows, self.cols, aug_rref[:, self.cols:])
-    #def _eval_transpose(self):
-    #    return MatrixBase.transpose(self)
-    #def _eval_conjugate(self):
-    #    return MatrixBase.conjugate(self)
+    def _eval_inverse(self, method=None, **kwargs):
+        return self.inv(method, **kwargs)
     def _eval_add(self, other):
         # we assume both arguments are dense matrices since
         # sparse matrices have a higher priority
@@ -448,243 +436,6 @@ class DenseMatrix(CommonMatrix, MatrixBase):
         # use iterated squaring to compute the power
         b = MatrixBase._exp_by_squaring(b, a, n)
         return self._new(b.rows, b.cols, b._mat, copy=False)
-    def _eval_pow(self, n):
-        from sympy.matrices import eye, diag, MutableMatrix
-        from sympy import binomial
-
-        def jordan_cell_power(jc, n):
-            N = jc.shape[0]
-            l = jc[0, 0]
-            for i in range(N):
-                    for j in range(N-i):
-                            bn = binomial(n, i)
-                            if isinstance(bn, binomial):
-                                    bn = bn._eval_expand_func()
-                            jc[j, i+j] = l**(n-i)*bn
-
-        P, jordan_cells = self.jordan_cells()
-        # Make sure jordan_cells matrices are mutable:
-        jordan_cells = [MutableMatrix(j) for j in jordan_cells]
-        for j in jordan_cells:
-            jordan_cell_power(j, n)
-        return self._new(P*diag(*jordan_cells)*P.inv())
-    #def _eval_cholesky(self):
-    #    return MatrixBase.cholesky(self)
-    #def _eval_LDLdecomposition(self):
-    #    return MatrixBase.LDLdecomposition(self)
-    #def _eval_lower_triangular_solve(self, other):
-    #    return MatrixBase.lower_triangular_solve(self, other)
-    #def _eval_upper_triangular_solve(self, other):
-    #    return MatrixBase.upper_triangular_solve(self, other)
-    def _eval_diagonal_solve(self, other):
-        return self._new(other.rows, other.cols, lambda i, j: other[i, j] / self[i, i])
-    def _eval_LUsolve(self, other, iszerofunc=_iszero):
-        A, perm = self.LUdecomposition_Simple(iszerofunc=_iszero)
-        n = self.rows
-        b = other.permuteFwd(perm).as_mutable()
-        # forward substitution, all diag entries are scaled to 1
-        for i in range(n):
-            for j in range(i):
-                scale = A[i, j]
-                b.zip_row_op(i, j, lambda x, y: x - y*scale)
-        # backward substitution
-        for i in range(n - 1, -1, -1):
-            for j in range(i + 1, n):
-                scale = A[i, j]
-                b.zip_row_op(i, j, lambda x, y: x - y*scale)
-            scale = A[i, i]
-            b.row_op(i, lambda x, _: x/scale)
-        return other._new(b)
-    def _eval_LUdecomposition(self, iszerofunc=_iszero):
-        combined, p = self.LUdecomposition_Simple(iszerofunc=_iszero)
-        L = InplaceMatrix.zeros(self.rows)
-        U = InplaceMatrix.zeros(self.rows)
-        for i in range(self.rows):
-            for j in range(self.rows):
-                if i > j:
-                    L[i, j] = combined[i, j]
-                else:
-                    if i == j:
-                        L[i, i] = 1
-                    U[i, j] = combined[i, j]
-        return self._new(L.rows, L.cols, L), self._new(U.rows, U.cols, U), p
-    def _eval_LUdecomposition_Simple(self, iszerofunc=_iszero):
-        n = self.rows
-        A = self.as_mutable()
-        p = []
-        # factorization
-        for j in range(n):
-            for i in range(j):
-                for k in range(i):
-                    A[i, j] = A[i, j] - A[i, k]*A[k, j]
-            pivot = -1
-            for i in range(j, n):
-                for k in range(j):
-                    A[i, j] = A[i, j] - A[i, k]*A[k, j]
-                # find the first non-zero pivot, includes any expression
-                if pivot == -1 and not iszerofunc(A[i, j]):
-                    pivot = i
-            if pivot < 0:
-                # this result is based on iszerofunc's analysis of the possible pivots, so even though
-                # the element may not be strictly zero, the supplied iszerofunc's evaluation gave True
-                raise ValueError("No nonzero pivot found; inversion failed.")
-            if pivot != j:  # row must be swapped
-                A.row_swap(pivot, j)
-                p.append([pivot, j])
-            scale = 1 / A[j, j]
-            for i in range(j + 1, n):
-                A[i, j] = A[i, j]*scale
-        return A, p
-    def _eval_LUdecompositionFF(self):
-        from sympy.matrices import SparseMatrix
-        zeros = SparseMatrix.zeros
-        eye = SparseMatrix.eye
-
-        n, m = self.rows, self.cols
-        U, L, P = self.as_mutable(), eye(n), eye(n)
-        DD = zeros(n, n)
-        oldpivot = 1
-
-        for k in range(n - 1):
-            if U[k, k] == 0:
-                for kpivot in range(k + 1, n):
-                    if U[kpivot, k]:
-                        break
-                else:
-                    raise ValueError("Matrix is not full rank")
-                U[k, k:], U[kpivot, k:] = U[kpivot, k:], U[k, k:]
-                L[k, :k], L[kpivot, :k] = L[kpivot, :k], L[k, :k]
-                P[k, :], P[kpivot, :] = P[kpivot, :], P[k, :]
-            L[k, k] = Ukk = U[k, k]
-            DD[k, k] = oldpivot*Ukk
-            for i in range(k + 1, n):
-                L[i, k] = Uik = U[i, k]
-                for j in range(k + 1, m):
-                    U[i, j] = (Ukk*U[i, j] - U[k, j]*Uik) / oldpivot
-                U[i, k] = 0
-            oldpivot = Ukk
-        DD[n - 1, n - 1] = oldpivot
-        return P, L, DD, U
-    def _eval_cofactorMatrix(self, method="berkowitz"):
-        out = self._new(self.rows, self.cols, lambda i, j:
-                self.cofactor(i, j, method))
-        return out
-    def _eval_minorMatrix(self, i, j):
-        M = self.as_mutable()
-        M.row_del(i)
-        M.col_del(j)
-        return self._new(M)
-    def _eval_QRdecomposition(self):
-        mat = self.as_mutable()
-
-        n = mat.rows
-        m = mat.cols
-        rank = n
-        row_reduced = mat.rref()[0]
-        for i in range(row_reduced.rows):
-            if row_reduced.row(i).norm() == 0:
-                rank -= 1
-        if not rank == mat.cols:
-            raise MatrixError("The rank of the matrix must match the columns")
-        Q, R = mat.zeros(n, m), mat.zeros(m)
-        for j in range(m):      # for each column vector
-            tmp = mat[:, j]     # take original v
-            for i in range(j):
-                # subtract the project of mat on new vector
-                tmp -= Q[:, i]*mat[:, j].dot(Q[:, i])
-                tmp.expand()
-            # normalize it
-            R[j, j] = tmp.norm()
-            Q[:, j] = tmp / R[j, j]
-            if Q[:, j].norm() != 1:
-                raise NotImplementedError(
-                    "Could not normalize the vector %d." % j)
-            for i in range(j):
-                R[i, j] = Q[:, i].dot(mat[:, j])
-        return self._new(Q), self._new(R)
-    def _eval_QRsolve(self, other):
-        Q, R = self.as_mutable().QRdecomposition()
-        y = Q.T*other
-        return self._new(MatrixBase._back_substitute(R, y))
-    def _eval_values(self):
-        return [i for i in self if not i.is_zero]
-    def _eval_rref(self, iszerofunc=_iszero, simplify=False):
-        simpfunc = simplify if isinstance(
-            simplify, FunctionType) else _simplify
-        # pivot: index of next row to contain a pivot
-        pivot, r = 0, self.as_mutable()
-        # pivotlist: indices of pivot variables (non-free)
-        pivotlist = []
-        for i in range(r.cols):
-            if pivot == r.rows:
-                break
-            if simplify:
-                r[pivot, i] = simpfunc(r[pivot, i])
-            if iszerofunc(r[pivot, i]):
-                for k in range(pivot, r.rows):
-                    if simplify and k > pivot:
-                        r[k, i] = simpfunc(r[k, i])
-                    if not iszerofunc(r[k, i]):
-                        r.row_swap(pivot, k)
-                        break
-                else:
-                    continue
-            scale = r[pivot, i]
-            r.row_op(pivot, lambda x, _: x / scale)
-            for j in range(r.rows):
-                if j == pivot:
-                    continue
-                scale = r[j, i]
-                r.zip_row_op(j, pivot, lambda x, y: x - scale*y)
-            pivotlist.append(i)
-            pivot += 1
-        return (self._new(r), pivotlist)
-    def _eval_diagonalize(self, reals_only=False, sort=False, normalize=False):
-        # if we've made it here, we're diagonalizable
-        if self._eigenvects is None:
-            self._eigenvects = self.eigenvects(simplify=True)
-        if sort:
-            self._eigenvects.sort(key=default_sort_key)
-            self._eigenvects.reverse()
-        diagvals = []
-        P = self._new(self.rows, 0, [])
-        for eigenval, multiplicity, vects in self._eigenvects:
-            for k in range(multiplicity):
-                diagvals.append(eigenval)
-                vec = vects[k]
-                if normalize:
-                    vec = vec / vec.norm()
-                P = P.col_insert(P.cols, vec)
-        D = diag(*diagvals)
-        self._diagonalize_clear_subproducts()
-        return (P, D)
-    def _eval_dual(self):
-        from sympy import LeviCivita
-        from sympy.matrices import zeros
-
-        M, n = self[:, :], self.rows
-        work = zeros(n)
-        if self.is_symmetric():
-            return work
-
-        for i in range(1, n):
-            for j in range(1, n):
-                acum = 0
-                for k in range(1, n):
-                    acum += LeviCivita(i, j, 0, k)*M[0, k]
-                work[i, j] = acum
-                work[j, i] = -acum
-
-        for l in range(1, n):
-            acum = 0
-            for a in range(1, n):
-                for b in range(1, n):
-                    acum += LeviCivita(0, l, a, b)*M[a, b]
-            acum /= 2
-            work[0, l] = -acum
-            work[l, 0] = acum
-
-        return work
     def _eval_row_join(self, other):
         a = InplaceMatrix.zeros(self.rows, self.cols + other.cols)
         a[:, :self.cols] = self
@@ -833,84 +584,6 @@ class DenseMatrix(CommonMatrix, MatrixBase):
             a.extend(self._mat[i::self.cols])
         return self._new(self.cols, self.rows, a)
 
-    def _eval_conjugate(self):
-        """By-element conjugation.
-
-        See Also
-        ========
-
-        transpose: Matrix transposition
-        H: Hermite conjugation
-        D: Dirac conjugation
-        """
-        return self.applyfunc(lambda x: x.conjugate())
-
-    def _eval_adjoint(self):
-        return self.T.C
-
-    def _eval_inverse(self, **kwargs):
-        """Return the matrix inverse using the method indicated (default
-        is Gauss elimination).
-
-        kwargs
-        ======
-
-        method : ('GE', 'LU', or 'ADJ')
-        iszerofunc
-        try_block_diag
-
-        Notes
-        =====
-
-        According to the ``method`` keyword, it calls the appropriate method:
-
-          GE .... inverse_GE(); default
-          LU .... inverse_LU()
-          ADJ ... inverse_ADJ()
-
-        According to the ``try_block_diag`` keyword, it will try to form block
-        diagonal matrices using the method get_diag_blocks(), invert these
-        individually, and then reconstruct the full inverse matrix.
-
-        Note, the GE and LU methods may require the matrix to be simplified
-        before it is inverted in order to properly detect zeros during
-        pivoting. In difficult cases a custom zero detection function can
-        be provided by setting the ``iszerosfunc`` argument to a function that
-        should return True if its argument is zero. The ADJ routine computes
-        the determinant and uses that to detect singular matrices in addition
-        to testing for zeros on the diagonal.
-
-        See Also
-        ========
-
-        inverse_LU
-        inverse_GE
-        inverse_ADJ
-        """
-        from sympy.matrices import diag
-
-        method = kwargs.get('method', 'GE')
-        iszerofunc = kwargs.get('iszerofunc', _iszero)
-        if kwargs.get('try_block_diag', False):
-            blocks = self.get_diag_blocks()
-            r = []
-            for block in blocks:
-                r.append(block.inv(method=method, iszerofunc=iszerofunc))
-            return diag(*r)
-
-        M = self.as_mutable()
-        if method == "GE":
-            rv = M.inverse_GE(iszerofunc=iszerofunc)
-        elif method == "LU":
-            rv = M.inverse_LU(iszerofunc=iszerofunc)
-        elif method == "ADJ":
-            rv = M.inverse_ADJ(iszerofunc=iszerofunc)
-        else:
-            # make sure to add an invertibility check (as in inverse_LU)
-            # if a new method is added.
-            raise ValueError("Inversion method unrecognized")
-        return self._new(rv)
-
     @classmethod
     def _eval_create_with_diag(cls, *entries):
         return diag(*entries)
@@ -978,60 +651,6 @@ class DenseMatrix(CommonMatrix, MatrixBase):
 
     def __ne__(self, other):
         return not self == other
-
-    def _eval_cholesky(self):
-        """Helper function of cholesky.
-        Without the error checks.
-        To be used privately. """
-        L = zeros(self.rows, self.rows)
-        for i in range(self.rows):
-            for j in range(i):
-                L[i, j] = (1 / L[j, j])*(self[i, j] -
-                    sum(L[i, k]*L[j, k] for k in range(j)))
-            L[i, i] = sqrt(self[i, i] -
-                    sum(L[i, k]**2 for k in range(i)))
-        return self._new(L)
-
-    def _eval_LDLdecomposition(self):
-        """Helper function of LDLdecomposition.
-        Without the error checks.
-        To be used privately.
-        """
-        D = zeros(self.rows, self.rows)
-        L = eye(self.rows)
-        for i in range(self.rows):
-            for j in range(i):
-                L[i, j] = (1 / D[j, j])*(self[i, j] - sum(
-                    L[i, k]*L[j, k]*D[k, k] for k in range(j)))
-            D[i, i] = self[i, i] - sum(L[i, k]**2*D[k, k]
-                for k in range(i))
-        return self._new(L), self._new(D)
-
-    def _eval_lower_triangular_solve(self, rhs):
-        """Helper function of function lower_triangular_solve.
-        Without the error checks.
-        To be used privately.
-        """
-        X = zeros(self.rows, rhs.cols)
-        for j in range(rhs.cols):
-            for i in range(self.rows):
-                if self[i, i] == 0:
-                    raise TypeError("Matrix must be non-singular.")
-                X[i, j] = (rhs[i, j] - sum(self[i, k]*X[k, j]
-                    for k in range(i))) / self[i, i]
-        return self._new(X)
-
-    def _eval_upper_triangular_solve(self, rhs):
-        """Helper function of function upper_triangular_solve.
-        Without the error checks, to be used privately. """
-        X = zeros(self.rows, rhs.cols)
-        for j in range(rhs.cols):
-            for i in reversed(range(self.rows)):
-                if self[i, i] == 0:
-                    raise ValueError("Matrix must be non-singular.")
-                X[i, j] = (rhs[i, j] - sum(self[i, k]*X[k, j]
-                    for k in range(i + 1, self.rows))) / self[i, i]
-        return self._new(X)
 
     def applyfunc(self, f):
         """Apply a function to each element of the matrix.
